backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anyio
attrs
bcrypt<4.0.0
beanie
black
boto3
botocore
certifi
cffi
charset-normalizer
click
cryptography
distro
dnspython
ecdsa
email-validator
fastapi
fastuuid
filelock
flake8
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
google_search_results
googleapis-common-protos
groq
grpcio
grpcio-status
h11
hf-xet
httpcore
httplib2
httpx
huggingface_hub
idna
importlib_metadata
iniconfig
isort
Jinja2
jiter
jmespath
jq
jsonschema
jsonschema-specifications
lazy-model
librt
litellm
loguru
markdown-it-py
MarkupSafe
mccabe
mdurl
motor
multidict
mypy
mypy_extensions
numpy
oauthlib
openai
packaging
pandas
passlib
pathspec
pillow
platformdirs
pluggy
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic-settings
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
PyMuPDF
pyparsing
pytest
python-dateutil
python-dotenv
python-jose
python-multipart
pytokens
PyYAML
referencing
regex
requests
requests-oauthlib
rich
rpds-py
rsa
s3transfer
s5cmd
shellingham
six
sniffio
starlette
stripe
tenacity
tiktoken
tokenizers
tqdm
typer
typer-slim
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchfiles
websockets
yarl
zipp
